Split migratableModel from migrator and give them their own directories
