+++ examples/hosts/app-integration/schema-upgrade/src/migrator/index.ts
@@ -1,0 +1,7 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+export * from "./interfaces";
+export * from "./migrator";
