+++ examples/hosts/app-integration/schema-upgrade/src/modelVersion2/containerModel.ts
@@ -7,7 +7,7 @@ import { TypedEventEmitter } from "@fluidframework/common-utils";
 import { AttachState, IContainer } from "@fluidframework/container-definitions";
 
 import { parseStringDataVersionTwo, readVersion } from "../dataTransform";
-import { MigrationState } from "../migrationInterfaces";
+import { MigrationState } from "../migratableModel";
 import type {
     IMigrationTool,
 } from "../migrationTool";
